always create ca.srl if not present

+++ create_ca_cert.sh
@@ -31,9 +31,6 @@ CA_SRL_FILE=${CA_SRL_FILE:-/ca/ca.srl}
 
 if [ -f "$CA_CRT_FILE" ] ; then
     logInfo "CA already exists. Good. We'll reuse it."
-    if [ ! -f "$CA_SRL_FILE" ] ; then
-        echo 01 > ${CA_SRL_FILE}
-    fi
 else
     logInfo "No CA was found. Generating one."
     logInfo "*** Please *** make sure to mount /ca as a volume -- if not, everytime this container starts, it will regenerate the CA and nothing will work."
@@ -56,8 +53,10 @@ EOF
     [[ ${DEBUG} -gt 0 ]] && logInfo "show the CA cert details"
     [[ ${DEBUG} -gt 0 ]] && openssl x509 -noout -text -in ${CA_CRT_FILE}
 
-    echo 01 > ${CA_SRL_FILE}
+fi
 
+if [ ! -f "$CA_SRL_FILE" ] ; then
+    echo 01 > ${CA_SRL_FILE}
 fi
 
 cd /certs
